lf compress: lfd loops: extract common wave run creation and spawn helpers
Consolidates duplicated code for creating wave runs and spawning execution tasks into a shared common.rs module.
