+++ rust/lfd/src/loops/loop_ticker.rs
@@ -3,9 +3,10 @@ use std::time::Duration;
 use tokio::task::JoinHandle;
 use tokio_util::sync::CancellationToken;
 
+use super::common::{create_wave_run_with_id, spawn_run_task_with_slot};
 use crate::executor::WaveExecutor;
 use crate::id::LfdId;
-use crate::proto::control::{StimulusKind, WaveRun, WaveRunStatus};
+use crate::proto::control::StimulusKind;
 use crate::scheduler::Scheduler;
 use crate::store::SharedStore;
 
@@ -83,7 +84,7 @@ async fn tick_loop_waves(
             continue;
         }
 
-        let run = match create_wave_run(store, &wave, &run_id) {
+        let run = match create_wave_run_with_id(store, &wave, &run_id) {
             Ok(run) => run,
             Err(err) => {
                 scheduler.release(run_id.as_str());
@@ -92,58 +93,6 @@ async fn tick_loop_waves(
             }
         };
 
-        let exec = executor.clone();
-        let store = store.clone();
-        let scheduler = scheduler.clone();
-        tokio::spawn(async move {
-            let run_id = LfdId::parse(&run.id).expect("run id should be valid");
-            if let Err(err) = exec.execute(&run_id).await {
-                tracing::error!(run_id = %run.id, error = %err, "run execution failed");
-                if let Ok(Some(mut run)) = store.get_wave_run(&run_id) {
-                    run.status = WaveRunStatus::WaveRunFailed as i32;
-                    run.error = Some(err.to_string());
-                    run.ended_at = Some(now_timestamp());
-                    let _ = store.update_wave_run(&run);
-                }
-            }
-            scheduler.release(&run.id);
-        });
-    }
-}
-
-fn create_wave_run(
-    store: &SharedStore,
-    wave: &crate::proto::control::Wave,
-    run_id: &LfdId,
-) -> anyhow::Result<WaveRun> {
-    let wave_id = LfdId::parse(&wave.id)?;
-    let last_run = store
-        .list_wave_runs(Some(&wave_id), Some(1))?
-        .into_iter()
-        .next();
-    let iteration = last_run.map(|run| run.iteration + 1).unwrap_or(0);
-
-    let run = WaveRun {
-        id: run_id.to_string(),
-        wave_id: wave.id.clone(),
-        iteration,
-        step_index: 0,
-        status: WaveRunStatus::WaveRunRunning as i32,
-        worktree: wave.repo.clone(),
-        branch: String::new(),
-        started_at: Some(now_timestamp()),
-        ended_at: None,
-        error: None,
-        flow_parents: Vec::new(),
-    };
-    store.create_wave_run(&run)?;
-    Ok(run)
-}
-
-fn now_timestamp() -> prost_types::Timestamp {
-    let now = time::OffsetDateTime::now_utc();
-    prost_types::Timestamp {
-        seconds: now.unix_timestamp(),
-        nanos: 0,
+        spawn_run_task_with_slot(store.clone(), executor.clone(), scheduler.clone(), run);
     }
 }
