+++ rust/lfd/src/loops/mod.rs
@@ -1,3 +1,4 @@
+mod common;
 mod cron;
 mod loop_ticker;
 mod recovery;
